Split TestFunctions.py in PointwiseFunctions.GeneralRelativity
