+++ tests/Unit/PointwiseFunctions/GeneralRelativity/Test_ComputeSpacetimeQuantities.cpp
@@ -21,14 +21,15 @@ namespace {
 template <size_t Dim, typename DataType>
 void test_compute_spacetime_metric(const DataType& used_for_size) {
   pypp::check_with_random_values<1>(
-      &gr::spacetime_metric<Dim, Frame::Inertial, DataType>, "TestFunctions",
-      "spacetime_metric", {{{-10., 10.}}}, used_for_size);
+      &gr::spacetime_metric<Dim, Frame::Inertial, DataType>,
+      "ComputeSpacetimeQuantities", "spacetime_metric", {{{-10., 10.}}},
+      used_for_size);
 }
 template <size_t Dim, typename DataType>
 void test_compute_inverse_spacetime_metric(const DataType& used_for_size) {
   pypp::check_with_random_values<1>(
       &gr::inverse_spacetime_metric<Dim, Frame::Inertial, DataType>,
-      "TestFunctions", "inverse_spacetime_metric", {{{-10., 10.}}},
+      "ComputeSpacetimeQuantities", "inverse_spacetime_metric", {{{-10., 10.}}},
       used_for_size);
 }
 template <size_t Dim, typename DataType>
@@ -36,14 +37,14 @@ void test_compute_derivatives_of_spacetime_metric(
     const DataType& used_for_size) {
   pypp::check_with_random_values<1>(
       &gr::derivatives_of_spacetime_metric<Dim, Frame::Inertial, DataType>,
-      "TestFunctions", "derivatives_of_spacetime_metric", {{{-10., 10.}}},
-      used_for_size);
+      "ComputeSpacetimeQuantities", "derivatives_of_spacetime_metric",
+      {{{-10., 10.}}}, used_for_size);
 }
 template <size_t Dim, typename DataType>
 void test_compute_spacetime_normal_vector(const DataType& used_for_size) {
   pypp::check_with_random_values<1>(
       &gr::spacetime_normal_vector<Dim, Frame::Inertial, DataType>,
-      "TestFunctions", "spacetime_normal_vector", {{{-10., 10.}}},
+      "ComputeSpacetimeQuantities", "spacetime_normal_vector", {{{-10., 10.}}},
       used_for_size);
 }
 template <size_t Dim, typename DataType>
@@ -64,8 +65,9 @@ void test_compute_spacetime_normal_one_form(const DataType& used_for_size) {
 template <size_t Dim, typename DataType>
 void test_compute_extrinsic_curvature(const DataType& used_for_size) {
   pypp::check_with_random_values<1>(
-      &gr::extrinsic_curvature<Dim, Frame::Inertial, DataType>, "TestFunctions",
-      "extrinsic_curvature", {{{-10., 10.}}}, used_for_size);
+      &gr::extrinsic_curvature<Dim, Frame::Inertial, DataType>,
+      "ComputeSpacetimeQuantities", "extrinsic_curvature", {{{-10., 10.}}},
+      used_for_size);
 }
 
 template <size_t Dim, typename T>
